Comp212-Solo-Project/src/requirements.txt
opencv-python
Pillow
customtkinter